perf: merge TapDrop and StarDrop
