+++ Assets/Prefab/Game/Tap_01.prefab
@@ -179,9 +179,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 7955322054321000439}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c57176cc05c23a743af16831f7ec76ba, type: 3}
+  m_Script: {fileID: 11500000, guid: 43a4a1be0a3c9af41918c882a121a7f8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _state: 0
   _startPos: 0
   _timing: 0
   _speed: 7
