+++ Assets/Script/Game/NoteControllers/NotePoolManager.cs
@@ -13,7 +13,6 @@ namespace MajdataPlay.Game
 
         NotePool<TapPoolingInfo, TapQueueInfo> tapPool;
         NotePool<HoldPoolingInfo, TapQueueInfo> holdPool;
-        SlideLauncherPool starPool;
         NotePool<TouchPoolingInfo, TouchQueueInfo> touchPool;
         NotePool<TouchHoldPoolingInfo, TouchQueueInfo> touchHoldPool;
         EachLinePool eachLinePool;
@@ -46,13 +45,13 @@ namespace MajdataPlay.Game
         {
             var tapParent = transform.GetChild(0);
             var holdParent = transform.GetChild(1);
-            var starParent = transform.GetChild(2);
+            //var starParent = transform.GetChild(2);
             var touchParent = transform.GetChild(4);
             var touchHoldParent = transform.GetChild(5);
             var eachLineParent = transform.GetChild(6);
             tapPool = new (tapPrefab, tapParent, tapInfos.ToArray(),128);
             holdPool = new (holdPrefab, holdParent, holdInfos.ToArray(),64);
-            starPool = new (starPrefab, starParent, starInfos.ToArray(),128);
+            //starPool = new (starPrefab, starParent, starInfos.ToArray(),128);
             touchPool = new (touchPrefab, touchParent, touchInfos.ToArray(),64);
             touchHoldPool = new (touchHoldPrefab, touchHoldParent, touchHoldInfos.ToArray(),64);
             eachLinePool = new (eachLinePrefab, eachLineParent, eachLineInfos.ToArray(),64);
@@ -69,17 +68,14 @@ namespace MajdataPlay.Game
             var currentSec = _gpManager.AudioTime;
             tapPool.Update(currentSec);
             holdPool.Update(currentSec);
-            starPool.Update(currentSec);
+            //starPool.Update(currentSec);
             touchPool.Update(currentSec);
             touchHoldPool.Update(currentSec);
             eachLinePool.Update(currentSec);
         }
         public void AddTap(TapPoolingInfo tapInfo)
         {
-            if (tapInfo.IsStar)
-                starInfos.Add(tapInfo);
-            else
-                tapInfos.Add(tapInfo);
+            tapInfos.Add(tapInfo);
         }
         public void AddHold(HoldPoolingInfo holdInfo)
         {
@@ -107,9 +103,6 @@ namespace MajdataPlay.Game
                 case HoldDrop hold:
                     holdPool.Collect(hold);
                     break;
-                case StarDrop star:
-                    starPool.Collect(star);
-                    break;
                 case EachLineDrop eachLine:
                     eachLinePool.Collect(eachLine);
                     break;
@@ -125,7 +118,6 @@ namespace MajdataPlay.Game
         {
             tapPool?.Destroy();
             holdPool?.Destroy();
-            starPool?.Destroy();
             touchPool?.Destroy();
             touchHoldPool?.Destroy();
             eachLinePool?.Destroy();
