--- Assets/Script/Game/Notes/StarDrop.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 309cdd5d8b4098f46aab9707812f563d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
