--- Assets/Script/Game/Notes/TapBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43a4a1be0a3c9af41918c882a121a7f8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
