--- Assets/Script/Misc/Types/Buffers/SlideLauncherPool.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4edf833e9e55ff34c912f0081197764f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
