+++ Assets/Script/Misc/Types/PoolingInfo/TapPoolingInfo.cs
@@ -10,12 +10,8 @@ namespace MajdataPlay.Types
     public sealed class TapPoolingInfo : NotePoolingInfo
     {
         public bool IsStar { get; init; }
-        public bool IsNoHead { get; init; }
         public bool IsDouble { get; init; }
-        public bool IsFakeStar { get; init; }
-        public bool IsForceRotate { get; init; }
         public float RotateSpeed { get; init; } = 1f;
         public TapQueueInfo QueueInfo { get; init; } = TapQueueInfo.Default;
-        public GameObject? Slide { get; init; } = null;
     }
 }
